Change base03 and green rgb color values
Green updated to experimental value used in
`altercation/vim-colors-solarized`: `#719e07`, equivalent to `L*a*b` 60
-30 60.

Base03 updated to `#002732`, changing `L*a*b` 15 -12 -12 to 13 -12 -12.
The new `L` value, decreased by 2, relates to the other base colors in
distances of 7, 32, 47 and 52. This reproduces the distances for the
light base colors: -5, -32, -47 and -52, plus increases the contrast
between base03 and base02.